refactor(types): centralize and reuse type definitions
- Moved common interfaces and types from multiple files:
  - Consolidated 'ProcessedTransactionData', 'Transaction', 'EventName', and others into 'src/libs/types.ts'.
  - Adjusted imports and updated relevant files:
    - 'groupedTransactions.ts' now imports Transaction-related types.
    - 'internal-notifications.ts' references 'NotificationType' and 'NotificationOptions' from the centralized types.
    - 'carousel/index.ts' relies on 'CalculateItemsPerSlideProps' from the consolidated types.
    - 'ipfs.ts' uses 'PinataHeaders', 'PinataResponse', 'IPFSMetadata' from the shared types file.
- Removed duplicate type definitions and improved maintainability.
