--- src/libs/components/carousel/types.ts
@@ -1,5 +1,0 @@
-export interface CalculateItemsPerSlideProps {
-  parentWidth: number;
-  minItemWidth: number;
-  maxItemWidth: number;
-}
